Level-7: Update ability to save to folders which have not been created
